will move cpp files into one directory
