+++ python/bvhar/_src/_ols.cpp
@@ -45,11 +45,11 @@ PYBIND11_MODULE(_ols, m) {
 
   py::class_<bvhar::MultiOls>(m, "MultiOls")
     .def(py::init<const Eigen::MatrixXd&, const Eigen::MatrixXd&>())
-		.def("estimateCoef", &bvhar::MultiOls::estimateCoef)
-		.def("fitObs", &bvhar::MultiOls::fitObs)
-		.def("estimateCov", &bvhar::MultiOls::estimateCov)
-		.def("fit", &bvhar::MultiOls::fit)
-    .def("return_ols_res", &bvhar::MultiOls::returnOlsRes);
+		// .def("estimateCoef", &bvhar::MultiOls::estimateCoef)
+		// .def("fitObs", &bvhar::MultiOls::fitObs)
+		// .def("estimateCov", &bvhar::MultiOls::estimateCov)
+		// .def("fit", &bvhar::MultiOls::fit)
+    .def("returnOlsRes", &bvhar::MultiOls::returnOlsRes);
 	
 	py::class_<bvhar::LltOls, bvhar::MultiOls>(m, "LltOls")
     .def(py::init<const Eigen::MatrixXd&, const Eigen::MatrixXd&>());
@@ -62,12 +62,12 @@ PYBIND11_MODULE(_ols, m) {
 			py::init<const Eigen::MatrixXd&, int, const bool, int>(),
 			py::arg("y"), py::arg("lag") = 1, py::arg("include_mean") = true, py::arg("method") = 1
 		)
-		.def("return_ols_res", &bvhar::OlsVar::returnOlsRes);
+		.def("returnOlsRes", &bvhar::OlsVar::returnOlsRes);
 	
 	py::class_<bvhar::OlsVhar>(m, "OlsVhar")
 		.def(
 			py::init<const Eigen::MatrixXd&, int, int, const bool, int>(),
 			py::arg("y"), py::arg("week") = 5, py::arg("month") = 22, py::arg("include_mean") = true, py::arg("method") = 1
 		)
-		.def("return_ols_res", &bvhar::OlsVhar::returnOlsRes);
+		.def("returnOlsRes", &bvhar::OlsVhar::returnOlsRes);
 }
